+++ CHANGES.rst
@@ -5,6 +5,8 @@ Development
 -----------
 
 * Change internal API to get access to all OAuth2 tokens from view
+* Change authorization request for offline credentials (to get a refresh_token)
+
 
 0.5.0 (2014-11-06)
 ------------------